refactored code involving reading/writing to submission directory to a new service SubmissionDirectoryService
