+++ backend/chromeApiServer/src/main/java/com/canvas/service/EvaluationService.java
@@ -26,21 +26,16 @@ import java.util.*;
  */
 @Service
 public class EvaluationService {
-    private final FileService fileService;
-    private final CanvasClientService canvasClientService;
-
-    private static final String MAKEFILE = "makefile";
+    private final SubmissionDirectoryService submissionDirectoryService;
 
     /**
-     * Constructor for creating our instances of FileService and CanvasClientService
+     * Constructor for creating the SubmissionDirectoryService instance
      *
-     * @param fileService         File service instance
-     * @param canvasClientService Canvas client service instance
+     * @param submissionDirectoryService SubmissionDirectoryService instance
      */
     @Autowired
-    public EvaluationService(FileService fileService, CanvasClientService canvasClientService) {
-        this.fileService = fileService;
-        this.canvasClientService = canvasClientService;
+    public EvaluationService(SubmissionDirectoryService submissionDirectoryService) {
+        this.submissionDirectoryService = submissionDirectoryService;
     }
 
     /**
@@ -52,131 +47,22 @@ public class EvaluationService {
      * @return return a CommandOutput response
      */
     public ResponseEntity<CommandOutput> compileStudentCodeFile(ExtensionUser user, MultipartFile[] files) throws CanvasAPIException {
-
-
         // Retrieve file json from Canvas
-        byte[] makefileBytes = new byte[0];
-        makefileBytes = canvasClientService.fetchFileUnderCourseAssignmentFolder(user, MAKEFILE);
-
-        // Write makefile to file
-        fileService.writeFileFromBytes(MAKEFILE, makefileBytes, user.getUserId());
+        submissionDirectoryService.writeMakefile(user, user.getUserId());
 
         // Write submitted code files
-        for (MultipartFile file : files) {
-            fileService.writeFileFromMultipart(file, user.getUserId());
-        }
+        submissionDirectoryService.writeSubmissionFiles(files, user.getUserId());
 
         // Compile the files and grab output
         CommandOutput commandOutput = compileCodeFiles(user.getUserId());
 
         // Cleanup
-        fileService.deleteDirectory(user.getUserId());
+        submissionDirectoryService.deleteDirectory(user.getUserId());
 
         // Generate response
         return new ResponseEntity<>(commandOutput, HttpStatus.OK);
     }
 
-    /**
-     * Method executes student submitted code for instructors/graders to grade.
-     *
-     * @param user User object for the extension
-     * @return return a CommandOutput response
-     */
-    public ResponseEntity<CommandOutput> executeCodeFile(ExtensionUser user) throws CanvasAPIException {
-        // Write makefile to directory
-        writeMakefile(user, user.getUserId());
-
-        // Fetch submission files from Canvas
-        Map<String, byte[]> submissionMap = canvasClientService.fetchStudentSubmissionFileBytes(user);
-
-        String userId = user.getUserId();
-
-        // Write submission files to directory
-        for (Map.Entry<String, byte[]> entry : submissionMap.entrySet()) {
-            // Key: filename, Value: file bytes to write
-            fileService.writeFileFromBytes(entry.getKey(), entry.getValue(), fileService.getFileDirectory(userId));
-        }
-
-        // If code compiles successfully
-        if (compileCodeFiles(userId).isSuccess()) {
-            // Execute the code
-            CommandOutput codeExecutionOutput = executeCodeFiles(userId);
-            System.out.println(codeExecutionOutput.getOutput());
-
-            // Cleanup
-            fileService.deleteDirectory(userId);
-
-            return new ResponseEntity<>(codeExecutionOutput, HttpStatus.OK);
-        } else {
-            // Cleanup
-            fileService.deleteDirectory(user.getStudentId());
-            return ResponseEntity.badRequest().build();
-        }
-    }
-
-    /**
-     * Generates the student submission and directory and writes the submission files to the directory.
-     *
-     * @param user Extension user
-     * @return Submission model response
-     * @throws CanvasAPIException
-     */
-    public ResponseEntity<Submission> generateSubmissionDirectory(ExtensionUser user) throws CanvasAPIException {
-        Submission submission = canvasClientService.fetchStudentSubmission(user);
-        String submissionDirectory = getDirectoryName(user.getCourseId(), user.getAssignmentId(), user.getStudentId());
-
-        writeMakefile(user, submissionDirectory);
-        SubmissionFile[] submissionFiles = writeSubmissionFiles(submission.getSubmissionFileBytes(), submissionDirectory);
-
-        submission.setSubmissionFiles(submissionFiles);
-        submission.setSubmissionDirectory(submissionDirectory);
-
-        return new ResponseEntity<>(submission, HttpStatus.OK);
-    }
-
-    /**
-     * Creates a unique directory name by hashing the arguments provided.
-     * @param idArgs IDs to hash
-     * @return unique hash value
-     */
-    private String getDirectoryName(String ... idArgs) {
-        String[] idList = Arrays.copyOf(idArgs, idArgs.length);
-        return String.valueOf(Arrays.hashCode(idList));
-    }
-
-    /**
-     * Helper method for writing submission files to the submission directory.
-     *
-     * @param submissionFilesBytes submission file byte map
-     * @param fileDirectory directory to write files to
-     * @return array of submission files
-     */
-    private SubmissionFile[] writeSubmissionFiles(Map<String, byte[]> submissionFilesBytes, String fileDirectory) {
-        List<SubmissionFile> submissionFiles = new ArrayList<>();
-
-        for (Map.Entry<String, byte[]> entry : submissionFilesBytes.entrySet()) {
-            String fileName = entry.getKey();
-            byte[] fileBytes = entry.getValue();
-            fileService.writeFileFromBytes(fileName, fileBytes, fileDirectory);
-            String[] fileContent = fileService.parseLinesFromFile(fileName, fileDirectory);
-            submissionFiles.add(new SubmissionFile(fileName, fileContent));
-        }
-
-        return submissionFiles.toArray(new SubmissionFile[0]);
-    }
-
-    /**
-     * Helper method for writing the makefile
-     *
-     * @param user User object the make file is associated with
-     * @throws CanvasAPIException
-     */
-    private void writeMakefile(ExtensionUser user, String directory) throws CanvasAPIException {
-        // Retrieve file json from Canvas
-        byte[] makefileBytes = canvasClientService.fetchFileUnderCourseAssignmentFolder(user, MAKEFILE);
-
-        fileService.writeFileFromBytes(MAKEFILE, makefileBytes, directory);
-    }
 
     /**
      * Helper method for compiling code files based on userId
@@ -193,17 +79,6 @@ public class EvaluationService {
     }
 
     /**
-     * Helper method for executing the code files based on userId
-     *
-     * @param userId canvas User ID for naming executable
-     * @return message associated with action
-     */
-    private CommandOutput executeCodeFiles(String userId) {
-        String exeFile = fileService.getFileNameWithExtension(".exe", userId);
-        return executeCommand(new String[]{exeFile}, userId);
-    }
-
-    /**
      * Helper method for executing process commands
      *
      * @param commands Array of commands to run for execution
@@ -211,7 +86,7 @@ public class EvaluationService {
      * @return message associated with action
      */
     private CommandOutput executeCommand(String[] commands, String userId) {
-        ProcessExecutor processExecutor = new ProcessExecutor(commands, fileService.getFileDirectory(userId));
+        ProcessExecutor processExecutor = new ProcessExecutor(commands, submissionDirectoryService.getSubmissionDirectory(userId));
         boolean compileSuccess = processExecutor.executeProcess();
         String output = processExecutor.getProcessOutput();
         return new CommandOutput(compileSuccess, output);
